Improved message for advanced mode.
We came to realize that "world" is perhaps neither inclusive enough nor
accurate enough for the purposes of this application. Although it is
traditional, and has a long standing place in the computer science
education community, we came to realize that further recognizing
humanities role in terms of its shepherding of the planet and its many
wonderful animals (all the fuzzy animals ^(o_O)^) is something we need
to be more conscious of as engaged Bereans.

Meh.